Frontend/requirements.txt
streamlit==1.31.0
pandas==2.1.4
plotly==5.18.0
scikit-learn==1.3.2
numpy==1.26.2
Pillow==10.2.0